refactor: split in multiple files
- use app provider to get per app locale bool
- check for per app locale before updating locale on changes
