--- lib/providers/android_info.dart
@@ -1,5 +1,0 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-import '../utilities/android_info.dart';
-
-final androidInfoProvider = FutureProvider((ref) => getAndroidInfo());
